moved addonHandler into a package
